requirements.txt
pandas==2.2.2
langchain==0.2.1
langchain_community==0.2.1
langchain-experimental==0.0.59
chainlit==1.1.202
pre-commit==3.7.1
ipykernel==6.29.3
llama_parse==0.4.3
llama_index==0.10.39
gpt4all==2.6.0
chromadb==0.5.0
unstructured==0.14.2
moviepy==1.0.3
assemblyai==0.26.0
jq==1.4.0
pdfminer.six==20231228
pillow_heif==0.16.0
opencv-python==4.9.0.80
pdf2image==1.17.0
unstructured_inference==0.7.33
pytesseract==0.3.10
unstructured_pytesseract==0.3.12
python-dotenv==1.0.1
autogen==1.0.16
pyautogen==0.2.2
ipython==8.25.0
markdownify==0.12.1
pypdf==4.2.0
sentence_transformers==3.0.0
langchain-openai==0.1.8
